Graph: Create EdgeSpec and PortSpec structs
Hopefully this will be more clear / useable than std::tuple :^)

+++ engine/src/Board/Parser.cc
@@ -159,7 +159,9 @@ static Graph from_stream(std::istream& char_stream)
                 if (neighbor.num == -1) {
                     continue;
                 }
-                edges.push_back({ token.num, neighbor.num, direction });
+                edges.push_back({ static_cast<size_t>(token.num),
+                                  static_cast<size_t>(neighbor.num),
+                                  direction });
 
                 // Build up the ports
                 if (token.ch == 'p') {
@@ -182,7 +184,9 @@ static Graph from_stream(std::istream& char_stream)
                     if (token.num > second_neighbor.num) {
                         continue;
                     }
-                    ports.push_back({ token.num, second_neighbor.num, get_orientation(direction) });
+                    ports.push_back({ static_cast<size_t>(token.num),
+                                      static_cast<size_t>(second_neighbor.num),
+                                      get_orientation(direction) });
                 }
             }
         }
